+++ vsce/src/core/workflows/initCodeBase.ts
@@ -1,6 +1,6 @@
 import * as vscode from "vscode";
 import * as wasm from "../../../pkg/neatcoder";
-import { AppStateManager } from "../appData";
+import { appDataManager } from "../appData";
 import { addLanguage } from "../commands/addLanguage";
 import { startLoading, stopLoading } from "../../utils/statusBar";
 
@@ -9,15 +9,15 @@ import { startLoading, stopLoading } from "../../utils/statusBar";
  * It goes through the following steps:
  * 1. Adds a language through the `addLanguage` function.
  * 2. Prompts the user to input their project description.
- * 3. If user input is received, it starts a loading indicator, triggers the start of a prompt process through the AppStateManager, and stops the loading indicator once completed.
+ * 3. If user input is received, it starts a loading indicator, triggers the start of a prompt process through the appDataManager, and stops the loading indicator once completed.
  *
  * @param llmParams - Parameters configuring the OpenAI client for task processing.
- * @param appManager - The AppStateManager instance responsible for managing the application state, including starting prompts.
+ * @param appManager - The appDataManager instance responsible for managing the application state, including starting prompts.
  * @returns Promise<void> - A promise that resolves once the prompt process has initiated and the user has been notified.
  */
 export async function initCodeBase(
   llmParams: wasm.OpenAIParams,
-  appManager: AppStateManager
+  appManager: appDataManager
 ): Promise<void> {
   {
     try {